Add support for Honest Profiler
This commit adds support for an additional profiler (Honest Profiler) instead of just JFR.
For this, it adds a `Profiler` abstraction, and now the `--profile` option requires to tell
which profiler to use. Honest Profiler requires either the `HP_HOME_DIR` environment variable
to be set, or the `hp-home` option to be set.

This does *not* add support for flame graph generation yet.