Don't run clean before benchmark
It is inconsistent with the --profile option, makes
the turnaround for benchmarking incremental changes slow
and breaks on builds that don't have a 'clean' task.

If there really is some leftover state affecting the measurement
then the user can run 'clean' manually. Instead, it is probably better
to fix the build, as 'clean' should not be necessary for correct results.