Replace jfr-flamegraph with official JMC library
The official library works on Java 7 and above,
can load recordings created by Java 7 and above.
As a result, we can now create flame graphs for all
Java versions.

The generation is also more efficient,
as we no longer call an external tool that loads the
data into memory every time, but instead load the data
once and reuse it.