Add mutation support for Android Layout files
Changes to Android layout files are a common situation to profile, especially when using Android Data binding, which
will trigger its own compiler. This PR adds support for this form of mutations to the profiler. It supports mutating
both normal layout files and data binding layouts, which have a special format.

This is done by inserting a non-visible view into the bottom of the layout with a random ID. The ID is important
as it guanartees that the API of the data binding model for this layout will also change.

The one thing not supported currently is mutating a layout file with a single element inside it, as in that case there
is nowhere for a fake element to be inserted. Likely the solution to this is to make the fake element a wrapper layout,
but I don't see much use for that use case right now, so I left it out.

Signed-off-by: Kyle Koser <kkoser@linkedin.com>