Fix the PidCollector to work with the configuration cache
by not relying on gradle.buildFinished
simply writing the file instead

Before it was using a separate i/o thread, joined when the build is
finished to write the pid to a file, only to prevent i/o to interfere
with the performance.

Now it writes it as soon as it can in a simpler way.

Even though the AsyncWriter usage make sense for e.g.
GradleTracingPlugin which is writing quite a lot of stuff,
it shouldn't change much for writing the pid number.

This could be done with a BuildService listener but there isn't any way
to get ProcessEnvironment in one yet so the pid would have to be
captured at configuration time anyway.

Signed-off-by: Paul Merlin <paul@gradle.com>