Refactor root project into 3 parts
- profiler-api => minimum classes needed by all profilers
- profiler-asyncprofiler => implementation for async profiler support
- everything else => remainder of the classes

This is an attempt to detangle the root project into separate
implementations for each supported profiler and build tool, eventually.

This also changes the way profilers are discovered by using a ServiceLoader.
This was necessary to avoid a cycle between profiler-api and the
implementations through ProfilerFactory.

No other behavioral changes should have been introduced.