Accept compressed base64 encoded data
To embed this into gradle-profile, we will need to accept embedded -stacks files.
This is because when opening an html file from a file:/// url, CORS does not allow it
to access other files. So, we must embed the data in the file.

To do this most efficiently, we compress the data, then we encode it in base64 to stick with printable characters and since base64 is well supported.
Then, we encode our demo data in this format, and subsequently decode the encoded demo data, proving the encoding system works.